requirements.txt
aiohttp==3.8.0
cycler==0.11.0
discord.py==2.0.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2